+++ src/Ninject.Extensions.Interception/Infrastructure/Language/ExtensionsForIBindingSyntax.cs
@@ -12,9 +12,12 @@
 
 #region Using Directives
 
+using System;
+using System.Reflection;
 using Ninject.Extensions.Interception.Advice;
 using Ninject.Extensions.Interception.Advice.Builders;
 using Ninject.Extensions.Interception.Advice.Syntax;
+using Ninject.Extensions.Interception.Parameters;
 using Ninject.Extensions.Interception.Registry;
 using Ninject.Syntax;
 
@@ -32,81 +35,40 @@ namespace Ninject.Extensions.Interception.Infrastructure.Language
         /// </summary>
         /// <typeparam name="T">The type associated with this binding.</typeparam>
         /// <param name="bindingSyntax">The binding syntax target.</param>
+        /// <param name="additionalInterfaces">The additional interfaces for the proxy.</param>
         /// <returns>
         ///     An <see cref="IAdviceTargetSyntax"/> instance which allows the attachment of an <see cref="IInterceptor"/>.
         /// </returns>
-        public static IAdviceTargetSyntax Intercept<T>( this IBindingWhenInNamedWithOrOnSyntax<T> bindingSyntax )
+        public static IAdviceTargetSyntax Intercept<T>( this IBindingOnSyntax<T> bindingSyntax, params Type[] additionalInterfaces )
         {
-            return DoIntercept( bindingSyntax );
+            return Intercept(bindingSyntax, 
+                mi => mi.DeclaringType != typeof(object), 
+                additionalInterfaces);
         }
 
         /// <summary>
         /// Indicates that instances associated with this binding will be proxied.
+        /// Only methods that match the specified predicate will be intercepted.
         /// </summary>
         /// <typeparam name="T">The type associated with this binding.</typeparam>
         /// <param name="bindingSyntax">The binding syntax target.</param>
-        /// <returns>
-        ///     An <see cref="IAdviceTargetSyntax"/> instance which allows the attachment of an <see cref="IInterceptor"/>.
-        /// </returns>
-        public static IAdviceTargetSyntax Intercept<T>( this IBindingInNamedWithOrOnSyntax<T> bindingSyntax )
+        /// <param name="methodPredicate">The method predicate that defines if a method shall be intercepted.</param>
+        /// <param name="additionalInterfaces">The additional interfaces for the proxy.</param>
+        /// <returns>An <see cref="IAdviceTargetSyntax" /> instance which allows the attachment of an <see cref="IInterceptor" />.</returns>
+        public static IAdviceTargetSyntax Intercept<T>(this IBindingOnSyntax<T> bindingSyntax, Predicate<MethodInfo> methodPredicate, params Type[] additionalInterfaces)
         {
-            return DoIntercept( bindingSyntax );
-        }
+            IKernel kernel = bindingSyntax.Kernel;
 
-        /// <summary>
-        /// Indicates that instances associated with this binding will be proxied.
-        /// </summary>
-        /// <typeparam name="T">The type associated with this binding.</typeparam>
-        /// <param name="bindingSyntax">The binding syntax target.</param>
-        /// <returns>
-        ///     An <see cref="IAdviceTargetSyntax"/> instance which allows the attachment of an <see cref="IInterceptor"/>.
-        /// </returns>
-        public static IAdviceTargetSyntax Intercept<T>( this IBindingNamedWithOrOnSyntax<T> bindingSyntax )
-        {
-            return DoIntercept( bindingSyntax );
-        }
+            foreach (var additionalInterface in additionalInterfaces)
+            {
+                bindingSyntax.BindingConfiguration.Parameters.Add(new AdditionalInterfaceParameter(additionalInterface));
+            }
 
-        /// <summary>
-        /// Indicates that instances associated with this binding will be proxied.
-        /// </summary>
-        /// <typeparam name="T">The type associated with this binding.</typeparam>
-        /// <param name="bindingSyntax">The binding syntax target.</param>
-        /// <returns>
-        ///     An <see cref="IAdviceTargetSyntax"/> instance which allows the attachment of an <see cref="IInterceptor"/>.
-        /// </returns>
-        public static IAdviceTargetSyntax Intercept<T>( this IBindingWithOrOnSyntax<T> bindingSyntax )
-        {
-            return DoIntercept( bindingSyntax );
-        }
-
-        /// <summary>
-        /// Indicates that instances associated with this binding will be proxied.
-        /// </summary>
-        /// <typeparam name="T">The type associated with this binding.</typeparam>
-        /// <param name="bindingSyntax">The binding syntax target.</param>
-        /// <returns>
-        ///     An <see cref="IAdviceTargetSyntax"/> instance which allows the attachment of an <see cref="IInterceptor"/>.
-        /// </returns>
-        public static IAdviceTargetSyntax Intercept<T>(this IBindingOnSyntax<T> bindingSyntax)
-        {
-            return DoIntercept(bindingSyntax);
-        }
-
-        /// <summary>
-        /// Constructs the interception advice to trigger according to the binding used.
-        /// </summary>
-        /// <param name="binding">The binding.</param>
-        /// <returns>
-        ///     An <see cref="IAdviceTargetSyntax"/> instance which allows the attachment of an <see cref="IInterceptor"/>.
-        /// </returns>
-        private static IAdviceTargetSyntax DoIntercept( IBindingSyntax binding )
-        {
-            IKernel kernel = binding.Kernel;
             IAdvice advice = kernel.Components.Get<IAdviceFactory>()
-                .Create(context => ReferenceEquals(binding.BindingConfiguration, context.Binding.BindingConfiguration));
-            kernel.Components.Get<IAdviceRegistry>().Register( advice );
+                .Create(context => ReferenceEquals(bindingSyntax.BindingConfiguration, context.Binding.BindingConfiguration), methodPredicate);
+            kernel.Components.Get<IAdviceRegistry>().Register(advice);
 
-            return new AdviceBuilder( advice );
+            return new AdviceBuilder(advice);
         }
     }
 }